requirements.txt
Flask>=3.0
pandas>=2.0
Pillow
requests
pyzbar; platform_system!="Windows"  # optional, requires libzbar
